Put available_drills and its tests in a different file
