+++ Makefile.am
@@ -5,6 +5,7 @@ bin_PROGRAMS = pcb2gcode
 pcb2gcode_SOURCES = \
     autoleveller.hpp \
     autoleveller.cpp \
+    available_drills.hpp \
     board.hpp \
     board.cpp \
     bg_helpers.hpp \
@@ -57,11 +58,13 @@ LIBS = $(glibmm_LIBS) $(gdkmm_LIBS) $(gerbv_LIBS) $(BOOST_PROGRAM_OPTIONS_LIBS) 
 
 EXTRA_DIST = millproject
 
-check_PROGRAMS = voronoi_tests eulerian_paths_tests segmentize_tests tsp_solver_tests units_tests
+check_PROGRAMS = voronoi_tests eulerian_paths_tests segmentize_tests tsp_solver_tests units_tests \
+                 available_drills_tests
 voronoi_tests_SOURCES = voronoi.hpp voronoi.cpp voronoi_tests.cpp
 eulerian_paths_tests_SOURCES = eulerian_paths_tests.cpp
 segmentize_tests_SOURCES = segmentize_tests.cpp
 tsp_solver_tests_SOURCES = tsp_solver_tests.cpp tsp_solver.hpp
 units_tests_SOURCES = units_tests.cpp units.hpp
+available_drills_tests_SOURCES = units_tests.cpp units.hpp
 
 TESTS = $(check_PROGRAMS)
